"enhancement": Add Save/add
Add save & add new button to existing records.

This would make sense if you just had to save for being able to add images/files. But I think in general it is a improvement, sometimes you have to edit something minor and want to be able to create a new record asap.