Added save button to new records
I really think we should add a save button to new records. Because in ModelAdmin sometimes I have has_many or many_many that only show up after save, so its bit annoying that I only have the choice to save and close or save and add new.

Besides some users would like to save and check if they filled in everything perfect again and if it saved well.